Rename test/__setup to test/index
