+++ fluent-dom/test/extra_text_markup_test.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import translateElement from '../src/overlay';
-import {elem} from './util';
+import {elem} from './index';
 
 suite('Localized text markup', function() {
   test('allowed element', function() {
