--- fluent-dom/test/util.js
@@ -1,7 +1,0 @@
-export function elem(name) {
-  return function(str) {
-    const element = document.createElement(name);
-    element.innerHTML = str;
-    return element;
-  }
-}
