--- fluent/test/__setup.js
@@ -1,3 +1,0 @@
-'use strict';
-
-require('../../fluent-intl-polyfill/src');
