+++ fluent/test/index.js
@@ -1,9 +1,12 @@
-import fs from "fs";
+'use strict';
 
-export function readfile(path) {
+const fs = require('fs');
+require('../../fluent-intl-polyfill/src');
+
+exports.readfile = function readfile(path) {
   return new Promise(function(resolve, reject) {
     fs.readFile(path, function(err, file) {
       return err ? reject(err) : resolve(file.toString());
     });
   });
-}
+};
