+++ fluent/test/parser_structure_test.js
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import { join } from 'path';
 import { readdir } from 'fs';
-import { readfile } from './util';
+import { readfile } from './index';
 
 import parse from '../src/parser';
